# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Scry API
  description: |
    API for Scry, a spaced repetition learning system with AI-generated cards.
    This API allows users to manage memos, cards, and review cards using a spaced repetition algorithm.
  version: 1.0.0
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        trace_id:
          type: string
          description: Unique ID for tracing the request through the system
      required:
        - error
...
    CardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique ID of the card
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns the card
        memo_id:
          type: string
          format: uuid
          description: ID of the memo this card was generated from
        content:
          type: object
          description: JSON content of the card (structure depends on card type)
        created_at:
          type: string
          format: date-time
          description: Time when the card was created
        updated_at:
          type: string
          format: date-time
          description: Time when the card was last updated
      required:
        - id
        - user_id
        - memo_id
        - content
        - created_at
        - updated_at
    EditCardRequest:
      type: object
      properties:
        content:
          type: object
          description: Updated JSON content for the card
      required:
        - content
    SubmitAnswerRequest:
      type: object
      properties:
        outcome:
          type: string
          enum: [again, hard, good, easy]
          description: The result of the card review
      required:
        - outcome
...
    UserCardStatsResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: ID of the user
        card_id:
          type: string
          format: uuid
          description: ID of the card
        interval:
          type: integer
          description: Current interval in days
        ease_factor:
          type: number
          format: float
          description: Current ease factor
        consecutive_correct:
          type: integer
          description: Count of consecutive correct answers
        last_reviewed_at:
          type: string
          format: date-time
          description: Time when the card was last reviewed
        next_review_at:
          type: string
          format: date-time
          description: Time when the card should next be reviewed
        review_count:
          type: integer
          description: Total number of reviews for this card
      required:
        - user_id
        - card_id
        - interval
        - ease_factor
        - consecutive_correct
        - last_reviewed_at
        - next_review_at
        - review_count
paths:
...
  /cards/next:
    get:
      tags:
        - Card Review
      summary: Get next card for review
      description: Retrieves the next card due for review
      operationId: getNextCard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Next card retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '204':
          description: No cards due for review
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cards/{id}/answer:
    post:
      tags:
        - Card Review
      summary: Submit answer for a card
      description: Submits the result of reviewing a card and updates its review schedule
      operationId: submitAnswer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Card ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCardStatsResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cards/{id}:
    put:
      tags:
        - Card Management
      summary: Edit a card
      description: Updates the content of an existing card
      operationId: editCard
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Card ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditCardRequest'
      responses:
        '204':
          description: Card updated successfully
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (card belongs to another user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Card Management
      summary: Delete a card
      description: |
        Deletes a card and its associated user_card_stats records.
        This operation is permanent and cannot be undone.
      operationId: deleteCard
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Card ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (card belongs to another user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...